# Nebula.API/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Nebula.API",
    "version": "1.0"
  },
  "paths": {
...
    "/watersheds/{watershedID}": {
      "get": {
        "tags": [
          "Watershed"
        ],
        "parameters": [
          {
            "name": "watershedID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WatershedDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatershedDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatershedDto"
                }
              }
            }
          }
        }
      }
    },
    "/watersheds/getBoundingBox": {
      "post": {
        "tags": [
          "Watershed"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatershedIDListDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/WatershedIDListDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/WatershedIDListDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BoundingBoxDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoundingBoxDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoundingBoxDto"
                }
              }
            }
          }
        }
      }
    },
    "/watersheds/{watershedName}/get-watershed-mask": {
      "get": {
        "tags": [
          "Watershed"
        ],
        "parameters": [
          {
            "name": "watershedName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BoundingBoxDto": {
        "type": "object",
        "additionalProperties": false
      },
...
      "WatershedDto": {
        "type": "object",
        "properties": {
          "WatershedID": {
            "type": "integer",
            "format": "int32"
          },
          "WatershedName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "WatershedIDListDto": {
        "type": "object",
        "properties": {
          "WatershedIDs": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}